--- test/ILLink.RoslynAnalyzer.Tests/Verifiers/CSharpCodeFixVerifier`2+Test.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.CSharp.Testing;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Testing.Verifiers;
-using Microsoft.CodeAnalysis.Text;
-
-namespace ILLink.RoslynAnalyzer.Tests
-{
-	public partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
-		where TAnalyzer : DiagnosticAnalyzer, new()
-		where TCodeFix : CodeFixProvider, new()
-	{
-		public class Test : CSharpCodeFixTest<TAnalyzer, TCodeFix, XUnitVerifier>
-		{
-			public Test ()
-			{
-				SolutionTransforms.Add ((solution, projectId) => {
-					var compilationOptions = solution.GetProject (projectId)!.CompilationOptions;
-					compilationOptions = compilationOptions!.WithSpecificDiagnosticOptions (
-						compilationOptions.SpecificDiagnosticOptions.SetItems (CSharpVerifierHelper.NullableWarnings));
-					solution = solution.WithProjectCompilationOptions (projectId, compilationOptions);
-
-					return solution;
-				});
-			}
-		}
-	}
-}
